requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-socketio>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
eventlet>=0.35
gunicorn>=21.2
apscheduler>=3.10
resend>=2.0
fpdf2>=2.7
authlib>=1.3
openpyxl>=3.1
